pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-revisor"
version = "0.1.0"
description = "Local shared-PC Bitbucket Cloud PR reviewer service"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
]

[tool.setuptools.packages.find]
include = ["service*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
